simple-website/requirements.txt
flask
google-generativeai
python-dotenv
markdown
datetime